Make gui generator classpath work
This is not exactly clean! We need a separate local copy of the Metawidget annotation
jar, since it contains MetawidgetAnnotationInspector which extends a class which is
in the Metawidget core. The annotations now appear twice, once in their own bundle
for no-dependency access from elsewhere, and once in the generator bundle to allow
the inspector to load. To make it work properly also required switching to an
Import-Package declaration for the annotations instead of Require-Bundle.